Moved simulator to backend
