--- Simulator/Dockerfile
@@ -1,15 +1,0 @@
-
-FROM python:3.13-slim
-
-WORKDIR /app
-
-COPY requirements.txt /app/
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-COPY Datasets /app/Datasets
-COPY . /app
-
-EXPOSE 8100
-
-CMD ["python3", "-m", "main"]
