--- Simulator/requirements.txt
@@ -1,3 +1,0 @@
-pandas
-psycopg2-binary
-numpy
